Use uint32/uint64 types (more semantically correct)
git-svn-id: http://chiral.j4cbo.com/svn/trunk@76 a827fe29-2235-0410-a7a9-cacdfe24a5d3

+++ chiral/os/epoll.py
@@ -23,8 +23,8 @@ class _epoll_data(ctypes.Union):
 	_fields_ = [
 		("ptr", ctypes.c_void_p),
 		("fd", ctypes.c_ulong),
-		("u32", ctypes.c_ulong),
-		("u64", ctypes.c_ulonglong)
+		("u32", ctypes.c_uint32),
+		("u64", ctypes.c_uint64)
 	]
 
 class _epoll_event(ctypes.Structure):
